Update components and styles for pages.
