+++ components/index.js
@@ -4,5 +4,4 @@ export * from './linkList'
 export Logo from './logo'
 export MarkdownElement from './MarkdownElement'
 export SliceHeader from './sliceHeader'
-export TechStack from './techStack'
 export * from './slices'
