+++ components/logo.jsx
@@ -1,16 +1,18 @@
 import React from 'react'
-import {Image} from './styled'
+import {Section, Image} from './styled'
 
 const Logo = ({src, customCss}) => (
-  <Image
-    src={src}
-    customCss={{
-      width: '6rem',
-      alignSelf: 'center',
-      marginBottom: '3rem',
-      ...customCss
-    }}
-  />
+  <Section>
+    <Image
+      src={src}
+      customCss={{
+        width: '6rem',
+        alignSelf: 'center',
+        marginBottom: '3rem',
+        ...customCss
+      }}
+    />
+  </Section>
 )
 
 Logo.displayName = 'Logo'
