+++ components/styled/link.js
@@ -1,12 +1,13 @@
 import React from 'react'
 import styled from 'react-emotion'
 import Link from 'next/link'
+import {Span} from './content'
 
 const A = styled('a')(({customCss}) => ({...customCss}))
 
 const NextLink = ({children, url, prefetch}) => (
   <Link href={url} passHref={true} prefetch={prefetch}>
-    {children}
+    <Span>{children}</Span>
   </Link>
 )
