+++ pages/cv/content.jsx
@@ -1,15 +1,14 @@
 import React, {Fragment} from 'react'
 import {Section} from '../../components'
 
-const Content = ({slices}) =>
-  console.log({slices}) || (
-    <Section>
-      {slices.map(({Component, data}, key) => (
-        <Fragment key={`slice-${key}`}>
-          <Component {...data} />
-        </Fragment>
-      ))}
-    </Section>
-  )
+const Content = ({slices}) => (
+  <Section>
+    {slices.map(({Component, data}, key) => (
+      <Fragment key={`slice-${key}`}>
+        <Component {...data} />
+      </Fragment>
+    ))}
+  </Section>
+)
 
 export default Content
